% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateCoefficients.R
\name{beta_first_s_covariates}
\alias{beta_first_s_covariates}
\title{First covariates have an effect}
\usage{
beta_first_s_covariates(p, s, values = rep(1, s), ...)
}
\arguments{
\item{p}{length of the vector}

\item{s}{the number of vectors that are non-zero}

\item{values}{vector of length s that contains the
values of the coefficients}
}
\value{
function that returns a vector of covariates
        where only the first s coefficients are non-zero
}
\description{
Returns a function that itself return a
coefficient vector where on the first
s coefficients are non-zero. The values
they take are given by the parameter values
}
