% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateCoefficients.R
\name{beta_spread_covariates}
\alias{beta_spread_covariates}
\title{Non-zero coefficients are spread out 'equally'}
\usage{
beta_spread_covariates(
  p,
  s,
  indices = floor(seq(1, p, length.out = s)),
  values = rep(1, length(indices)),
  ...
)
}
\arguments{
\item{p}{length of the vector}

\item{s}{the number of vectors that are non-zero}

\item{indices}{the indices of the coefficients that are non-zero
(Default: equally spread out)}

\item{values}{vector of length s that contains the
values of the coefficients}
}
\value{
function that returns a vector of covariates
        where only the first s coefficients are non-zero
}
\description{
Returns a function that itself return a
coefficient vector where non-zero coefficients
are spread out 'equally'. The values
they take are given by the parameter values
}
