% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMatrices.R
\name{corrmat_block}
\alias{corrmat_block}
\title{Block Correlation Matrix}
\usage{
corrmat_block(p, rho, blocksize, ...)
}
\arguments{
\item{p}{matrix size}

\item{rho}{controls the correlation strength}

\item{blocksize}{the size of the blocks.}
}
\value{
function that returns a block matrix
}
\description{
Returns a function that itself returns a p x p
matrix with blocksize blocks.
Note that the blocksize should fit the matrix,
i.e., mod(p, blocksize) should be 0
}
