% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMatrices.R
\name{corrmat_identity}
\alias{corrmat_identity}
\title{Identity Correlation Matrix}
\usage{
corrmat_identity(p, ...)
}
\arguments{
\item{p}{matrix size}
}
\value{
function that returns an identity matrix
}
\description{
Returns a function that itself returns a p x p
identity matrix
}
