% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMatrices.R
\name{corrmat_toeplitz}
\alias{corrmat_toeplitz}
\title{Toeplitz Correlation Matrix}
\usage{
corrmat_toeplitz(p, rho, ...)
}
\arguments{
\item{p}{matrix size}

\item{rho}{controls the correlation strength}
}
\value{
function that returns a Toeplitz matrix
}
\description{
Returns a function that itself returns a p x p
Toeplitz matrix
}
