% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simsham}
\alias{simsham}
\title{Simulate normally distributed dataset}
\usage{
simsham(
  n = 10,
  p = 20,
  s = 5,
  corrmat = corrmat_identity(p),
  beta = beta_first_s_covariates(p, s),
  snr = 1
)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of parameters}

\item{s}{number of non-zero coefficients}

\item{snr}{signal-to-noise ratio}

\item{corr_matrix}{function that returns
the correlation matrix}

\item{beta_type}{function that returns the
beta coefficient vector}
}
\value{
A list with X, y
}
\description{
Returns a normally distributed dataset with a
correlation matrix that is given by the function
corr_matrix and the non-zero coefficients as given by
the function beta_type
}
